requirements.txt
aiosmtpd==1.4.5
alembic==1.13.1
atpublic==4.1.0
attrs==23.2.0
Babel==2.15.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
elastic-transport==8.13.1
elasticsearch==8.13.2
email_validator==2.1.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
furl==2.1.3
greenlet==3.0.3
guess-language-spirit==0.5.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
langdetect==1.0.9
Mako==1.3.3
MarkupSafe==2.1.5
orderedmultidict==1.0.1
packaging==24.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
repoze.lru==0.7
requests==2.32.3
Routes==2.5.1
six==1.16.0
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
typing_extensions==4.11.0
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.2
WTForms==3.1.2
